NODE-404 Report system information only at end, report only waves and metrics sections
